# src/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "maklaGo API",
    "version": "1.0.0",
    "description": "Restaurant POS backend. **Tenant**: optional header `x-tenant-id` (slug or registry UUID). With `Authorization: Bearer` staff JWT, the tenant is resolved from the token when the header is omitted. **Language**: optional header **`lang`** (`fr`, `en`, `ar`; default **`fr`**) on every request — error messages and localized strings follow it; response includes **`Content-Language`**. **`POST /api/v1/auth/login`** accepts optional `x-tenant-id` for tenant-scoped login; without it, the normalized email must exist in registry **`staff_login_directory`** (slugless directory login). **`GET /api/v1/orders`** returns enriched catalog labels on lines (dashboard); **`GET /api/v1/mobile/orders`** uses a slimmer shell than **`POST /orders`**: no **`staff`**, no nested **`table`** object; dine-in adds root **`tableId`** + **`tableNumber`**. Lines add **`products[].name`**, **`extras[].name`**, optional **`extras[].typeId`**, optional **`compositionSnapshot`**. **`POST /api/v1/orders`** accepts **`products`** snapshots (`categoryId`, `id`, `count`, **`price`**, **`extras`**) for analytics; **`subtotal`**, **`tva`**, **`total`** use major currency units in JSON (e.g. 2.5); DB keeps cents internally (must match server); new orders start as **`waiting`**. **Platform** routes use `x-platform-key`. Errors: `{ \"error\": { \"code\", \"message\", \"details?\" } }`."
  },
  "servers": [
    {
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Staff JWT from `/api/v1/auth/login`."
      },
      "PlatformKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-platform-key",
        "description": "Must match `PLATFORM_API_KEY` on the server."
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            }
          }
        }
      },
...
      "OrderStatus": {
        "type": "string",
        "enum": [
          "waiting",
          "confirmed",
          "preparing",
          "completed",
          "canceled"
        ]
      },
...
      "PatchOrderTableRequest": {
        "type": "object",
        "required": ["tableId"],
        "properties": {
          "tableId": {
            "type": "string",
            "minLength": 12,
            "maxLength": 36,
            "description": "Active restaurant table id for this dine-in order."
          }
        }
      },
...
      "OrderProductEnriched": {
        "type": "object",
        "description": "Backoffice order line: product/category labels and resolved extra names joined from catalog.",
        "properties": {
          "categoryId": { "type": "string", "minLength": 12, "maxLength": 36 },
          "id": { "type": "string", "minLength": 12, "maxLength": 36 },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "kind": { "type": "string", "enum": ["simple", "composed"] },
          "categoryName": { "type": "string" },
          "image": { "type": "string", "nullable": true },
          "count": { "type": "integer" },
          "price": { "type": "number", "description": "Saved base unit price for this line (major currency units)." },
          "extras": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string" },
                "count": { "type": "integer" },
                "price": { "type": "number" },
                "name": { "type": "string", "description": "Present when catalog extra exists." }
              }
            }
          },
          "compositionSnapshot": { "type": "object", "additionalProperties": true, "nullable": true }
        }
      },
      "OrderTableSerialized": {
        "type": "object",
        "description": "Table row embedded on dine-in orders. Display label is **`T` + `tableNumber`** (no separate name field). **`zone`** is null when the table has no zone.",
        "properties": {
          "id": { "type": "string", "minLength": 12, "maxLength": 36 },
          "tableNumber": { "type": "integer", "minimum": 1 },
          "seatCount": { "type": "integer", "minimum": 1 },
          "zoneId": { "type": "string", "minLength": 12, "maxLength": 36, "nullable": true },
          "zone": {
            "type": "object",
            "nullable": true,
            "properties": {
              "id": { "type": "string", "minLength": 12, "maxLength": 36 },
              "name": { "type": "string" },
              "sortOrder": { "type": "integer" }
            }
          },
          "sortOrder": { "type": "integer" },
          "status": { "type": "string", "enum": ["free", "occupied"] },
          "isActive": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "OrderSerializedEnriched": {
        "type": "object",
        "description": "Backoffice order (`GET /orders`, `GET /orders/{id}`, `PATCH …/status` HTTP body): enriched **`products`** (`OrderProductEnriched`). Shell matches **`OrderSerialized`** (table rules, `staff`, etc.).",
        "properties": {
          "id": { "type": "string", "minLength": 12, "maxLength": 36 },
          "reference": {
            "type": "string",
            "description": "Unique ticket reference: **8** lowercase hexadecimal characters for newly created orders.",
            "example": "a3f9c21d"
          },
          "commandNumber": { "type": "integer" },
          "commandDate": { "type": "string", "format": "date" },
          "status": { "$ref": "#/components/schemas/OrderStatus" },
          "orderType": { "type": "string", "enum": ["dine_in", "takeaway"] },
          "tableId": {
            "type": "string",
            "minLength": 12,
            "maxLength": 36,
            "description": "Present only for dine-in orders with an assigned table."
          },
          "staffId": { "type": "string", "minLength": 12, "maxLength": 36 },
          "note": { "type": "string", "nullable": true },
          "customerName": { "type": "string", "nullable": true },
          "discount": { "type": "number", "description": "Order-level discount percent (0–100)." },
          "discountPrice": { "type": "number" },
          "paymentMethod": { "type": "string", "enum": ["cash", "card", "unpaid"] },
          "subtotal": { "type": "number", "description": "Sum of line net totals (major currency units)." },
          "tva": { "type": "number", "description": "Sum of line TVA amounts (major currency units)." },
          "total": { "type": "number", "description": "Final total after discount (major currency units)." },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "products": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/OrderProductEnriched" }
          },
          "table": {
            "$ref": "#/components/schemas/OrderTableSerialized",
            "description": "Present only with `tableId` on dine-in orders when table row was loaded.",
            "nullable": true
          },
          "staff": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          }
        }
      },
      "TableBody": {
        "type": "object",
        "required": [
          "tableNumber"
        ],
        "properties": {
          "tableNumber": {
            "type": "integer",
            "minimum": 1,
            "description": "Numeric label; **unique together with `zoneId`** (different zones may reuse the same number)."
          },
          "zoneId": {
            "type": "string",
            "minLength": 12,
            "maxLength": 36,
            "description": "**POST /tables**: required — table numbers are unique **per zone**. **PATCH**: optional; omit keys you are not updating."
          },
          "seatCount": {
            "type": "integer",
            "minimum": 1,
            "maximum": 999,
            "default": 4
          },
          "sortOrder": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": ["free", "occupied"],
            "description": "Patch only; normally derived from open dine-in orders."
          }
        }
      },
...
    },
    "parameters": {
      "OrderId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 12,
          "maxLength": 36,
          "description": "Order public id (lowercase hex or legacy UUID)."
        }
      },
      "TableId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "TenantHeaderOptional": {
        "name": "x-tenant-id",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Tenant slug (e.g. demo) or registry tenant UUID. Optional when the request includes a valid staff Bearer JWT—the server resolves tenant from the token."
      },
...
    }
  },
  "paths": {
...
    "/api/v1/orders/{id}/table": {
      "patch": {
        "tags": ["Orders"],
        "summary": "Move dine-in order to another table",
        "description": "Updates **`tableId`** on an open dine-in order (`confirmed` or `preparing`). Refreshes occupancy on old and new tables. Broadcasts **`order.updated`** and **`table.updated`** on the staff WebSocket. Target table must not already have another **`confirmed`** or **`preparing`** dine-in order.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/TenantHeaderOptional" },
          { "$ref": "#/components/parameters/OrderId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PatchOrderTableRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "order": { "$ref": "#/components/schemas/OrderSerializedEnriched" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation / not dine-in / order closed / table not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "409": {
            "description": "Target table already occupied",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tables": {
      "get": {
        "tags": [
          "Tables"
        ],
        "summary": "List active tables",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantHeaderOptional"
          }
        ]
      },
      "post": {
        "tags": [
          "Tables"
        ],
        "summary": "Create table",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TableBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantHeaderOptional"
          }
        ]
      }
    },
    "/api/v1/tables/{id}/toggle-status": {
      "post": {
        "tags": ["Tables"],
        "summary": "Toggle table free / occupied",
        "description": "Sets **`status`** to the opposite of the current value and emits **`table.updated`** on the staff WebSocket.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/TenantHeaderOptional" },
          { "$ref": "#/components/parameters/TableId" }
        ],
        "responses": {
          "200": {
            "description": "OK — JSON `{ table, status }`"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/tables/{id}": {
      "patch": {
        "tags": [
          "Tables"
        ],
        "summary": "Update table",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantHeaderOptional"
          },
          {
            "$ref": "#/components/parameters/TableId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TableBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Tables"
        ],
        "summary": "Soft-delete table (isActive=false)",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/TenantHeaderOptional"
          },
          {
            "$ref": "#/components/parameters/TableId"
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  }
}